+++ src/stores/index.ts
@@ -1,2 +1,3 @@
 export {foodStore} from './foodStore.ts';
 export {recipeStore} from './recipeStore.ts';
+export {calendarStore} from './calendarStore.ts';
