+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export {hS, mS, vS} from './useScale.ts';
 export * from './uploadImg.ts';
 export * from './camera.ts';
 export * from './handleTotals.ts';
+export * from './calendarUtils.ts';
